Lambda function for generating windows
